pyproject.toml
[tool.poetry]
name = "chesscards"
version = "0.1.0"
description = "spaced repetition for chess tactics"
authors = ["oschlueter <10252511+oschlueter@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fsrs = "^2.0.0"
CairoSVG = "^2.0"
chess = "^1.0"
matplotlib = "^3.0"
pillow = "^10.0"
pandas = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0.0"
black = "^24.3.0"
freezegun = "^1.4.0"

[tool.pytest.ini_options]
addopts = [
    '--cov=chesscards',
]

[tool.coverage.report]
show_missing = true

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
